+++ model_navigator/commands/export/exporters/torch2exportedprogram.py
@@ -77,10 +77,8 @@ def export(
         max_batch_size = None
         expanded_sample = expand_sample(conversion_sample, input_metadata, batch_dim=batch_dim, batch_size=None)
     else:
-        # WAR for to big batch size value
-        max_batch_size = max_batch_size if max_batch_size < 2048 else max_batch_size - 1
         # WAR to make data dynamic
-        batch_size = min(2, max_batch_size)  # select the minimum value to expand samples
+        batch_size = 2 if max_batch_size > 1 else 1  # select the minimum value to expand samples
         expanded_sample = expand_sample(conversion_sample, input_metadata, batch_dim=batch_dim, batch_size=batch_size)
 
     dummy_input = {n: torch.from_numpy(val).to(target_device) for n, val in expanded_sample.items()}
@@ -103,15 +101,15 @@ def export(
         if not tensor_metadata:
             continue
 
-        dynamic_shapes_ = {}
+        dynamic_shape_map = {}
         if max_batch_size is not None and max_batch_size > 1 and len(tensor_metadata.shape) > 0:
-            dynamic_shapes_[0] = torch.export.Dim(f"{name}_batch", min=1, max=max_batch_size)
+            dynamic_shape_map[0] = torch.export.Dim(f"{name}_batch", min=1, max=max_batch_size)
 
         for idx in range(1, len(spec_.min)):
             if spec_.min[idx] != spec_.max[idx]:
-                dynamic_shapes_[idx] = torch.export.Dim(f"{name}__{idx}", min=spec_.min[idx], max=spec_.max[idx])
+                dynamic_shape_map[idx] = torch.export.Dim(f"{name}__{idx}", min=spec_.min[idx], max=spec_.max[idx])
 
-        dynamic_shapes.append(dynamic_shapes_)
+        dynamic_shapes.append(dynamic_shape_map)
 
     try:
         exported_model = torch.export.export(